+++ src/main/java/dandi/dandi/post/application/service/PostService.java
@@ -2,31 +2,15 @@ package dandi.dandi.post.application.service;
 
 import dandi.dandi.common.exception.ForbiddenException;
 import dandi.dandi.common.exception.NotFoundException;
-import dandi.dandi.post.application.port.in.FeedResponse;
-import dandi.dandi.post.application.port.in.LikedPostResponse;
-import dandi.dandi.post.application.port.in.LikedPostResponses;
-import dandi.dandi.post.application.port.in.MyPostByTemperatureResponse;
-import dandi.dandi.post.application.port.in.MyPostResponse;
-import dandi.dandi.post.application.port.in.MyPostResponses;
-import dandi.dandi.post.application.port.in.MyPostsByTemperatureResponses;
-import dandi.dandi.post.application.port.in.PostDetailResponse;
 import dandi.dandi.post.application.port.in.PostRegisterCommand;
 import dandi.dandi.post.application.port.in.PostRegisterResponse;
-import dandi.dandi.post.application.port.in.PostResponse;
 import dandi.dandi.post.application.port.in.PostUseCase;
-import dandi.dandi.post.application.port.in.PostWriterResponse;
 import dandi.dandi.post.application.port.out.PostPersistencePort;
 import dandi.dandi.post.application.port.out.PostReportPersistencePort;
 import dandi.dandi.post.domain.Post;
-import dandi.dandi.post.domain.TemperatureSearchCondition;
 import dandi.dandi.post.domain.Temperatures;
 import dandi.dandi.post.domain.WeatherFeeling;
-import dandi.dandi.postlike.application.port.out.PostLikePersistencePort;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,18 +18,14 @@ import org.springframework.transaction.annotation.Transactional;
 public class PostService implements PostUseCase {
 
     private static final int POST_IMAGE_URL_INDEX = 1;
-    private static final int TEMPERATURE_SEARCH_THRESHOLD = 3;
 
     private final PostPersistencePort postPersistencePort;
-    private final PostLikePersistencePort postLikePersistencePort;
     private final PostReportPersistencePort postReportPersistencePort;
     private final String imageAccessUrl;
 
-    public PostService(PostPersistencePort postPersistencePort, PostLikePersistencePort postLikePersistencePort,
-                       PostReportPersistencePort postReportPersistencePort,
+    public PostService(PostPersistencePort postPersistencePort, PostReportPersistencePort postReportPersistencePort,
                        @Value("${cloud.aws.cloud-front.uri}") String imageAccessUrl) {
         this.postPersistencePort = postPersistencePort;
-        this.postLikePersistencePort = postLikePersistencePort;
         this.postReportPersistencePort = postReportPersistencePort;
         this.imageAccessUrl = imageAccessUrl;
     }
@@ -69,16 +49,6 @@ public class PostService implements PostUseCase {
     }
 
     @Override
-    @Transactional(readOnly = true)
-    public PostDetailResponse getPostDetails(Long memberId, Long postId) {
-        Post post = postPersistencePort.findById(postId)
-                .orElseThrow(NotFoundException::post);
-        boolean mine = post.isWrittenBy(memberId);
-        boolean liked = postLikePersistencePort.existsByPostIdAndMemberId(memberId, postId);
-        return new PostDetailResponse(post, mine, liked, imageAccessUrl);
-    }
-
-    @Override
     @Transactional
     public void deletePost(Long memberId, Long postId) {
         Post post = postPersistencePort.findById(postId)
@@ -91,64 +61,6 @@ public class PostService implements PostUseCase {
         if (!post.isWrittenBy(memberId)) {
             throw ForbiddenException.postDeletion();
         }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public MyPostResponses getMyPostIdsAndPostImageUrls(Long memberId, Pageable pageable) {
-        Slice<Post> posts = postPersistencePort.findByMemberId(memberId, pageable);
-        List<MyPostResponse> myPostResponses = posts.stream()
-                .map(post -> new MyPostResponse(post, imageAccessUrl))
-                .collect(Collectors.toUnmodifiableList());
-        return new MyPostResponses(myPostResponses, posts.isLast());
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public FeedResponse getPostsByTemperature(Long memberId, Double minTemperature, Double maxTemperature,
-                                              Pageable pageable) {
-        Temperatures temperatures = new Temperatures(minTemperature, maxTemperature);
-        TemperatureSearchCondition searchCondition =
-                temperatures.calculateTemperatureSearchCondition(TEMPERATURE_SEARCH_THRESHOLD);
-        Slice<Post> posts = postPersistencePort.findByTemperature(memberId, searchCondition, pageable);
-        return convertToFeedResponse(memberId, posts);
-    }
-
-    private FeedResponse convertToFeedResponse(Long memberId, Slice<Post> posts) {
-        List<PostResponse> postResponses = posts.stream()
-                .map(post -> new PostResponse(post, post.isLikedBy(memberId), imageAccessUrl))
-                .collect(Collectors.toUnmodifiableList());
-        return new FeedResponse(postResponses, posts.isLast());
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public MyPostsByTemperatureResponses getMyPostsByTemperature(Long memberId, Double minTemperature,
-                                                                 Double maxTemperature, Pageable pageable) {
-        Temperatures temperatures = new Temperatures(minTemperature, maxTemperature);
-        TemperatureSearchCondition searchCondition =
-                temperatures.calculateTemperatureSearchCondition(TEMPERATURE_SEARCH_THRESHOLD);
-        Slice<Post> myPostsByTemperature =
-                postPersistencePort.findByMemberIdAndTemperature(memberId, searchCondition, pageable);
-        return convertToMyPostsByTemperatureResponses(memberId, myPostsByTemperature);
-    }
-
-    private MyPostsByTemperatureResponses convertToMyPostsByTemperatureResponses(Long memberId,
-                                                                                 Slice<Post> myPostsByTemperature) {
-        List<MyPostByTemperatureResponse> myPostByTemperatureResponses = myPostsByTemperature.stream()
-                .map(post -> new MyPostByTemperatureResponse(post, post.isLikedBy(memberId), imageAccessUrl))
-                .collect(Collectors.toUnmodifiableList());
-        PostWriterResponse postWriterResponse = generatePostWriterResponse(myPostsByTemperature.getContent());
-        return new MyPostsByTemperatureResponses(
-                myPostByTemperatureResponses, postWriterResponse, myPostsByTemperature.isLast());
-    }
-
-    private PostWriterResponse generatePostWriterResponse(List<Post> myPostsByTemperature) {
-        if (myPostsByTemperature.isEmpty()) {
-            return null;
-        }
-        Post post = myPostsByTemperature.get(0);
-        return new PostWriterResponse(post, imageAccessUrl);
     }
 
     @Override
@@ -169,15 +81,5 @@ public class PostService implements PostUseCase {
         if (postReportPersistencePort.existsByMemberIdAndPostId(memberId, postId)) {
             throw new IllegalStateException("이미 신고한 게시글입니다.");
         }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public LikedPostResponses getLikedPost(Long memberId, Pageable pageable) {
-        Slice<Post> posts = postPersistencePort.findLikedPosts(memberId, pageable);
-        List<LikedPostResponse> likedPostResponses = posts.stream()
-                .map(post -> new LikedPostResponse(post, imageAccessUrl))
-                .collect(Collectors.toUnmodifiableList());
-        return new LikedPostResponses(likedPostResponses, posts.isLast());
     }
 }
